requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-telegram-bot==20.8
pydantic==2.5.2
sqlite-utils==3.35
